ebpf: implement PossibleCPU on Windows
Move the Linux implementation into internal/linux and add a wrapper
around the Windows GetMaximumProcessorCount function.
